dev-requirements.txt
ipython==7.19.0
pylint-django==2.3.0